% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/bagging_pp.R
\name{bagging_pp}
\alias{bagging_pp}
\title{List with bagging PP.tree error and bagging classification.}
\usage{
bagging_pp(data, boot, ...)
}
\arguments{
\item{data}{is a data frame with the data to classify. Class factor in the first column}

\item{boot}{object from bootstrap_pp (list of PP.Tree objets for bootstrap samples)}
}
\value{
list with the error, predicted classes and confusion matrix.
}
\description{
List with bagging PP.tree error and bagging classification.
}
\examples{
data1<-iris[,5:1]
output<-bootstrap_pp(data1,training,ntree=50,index="LDA")
bagging_pp(data1, output)
}

