% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootstrap}
\alias{bootstrap}
\title{Draws bootstrap samples with strata option.}
\usage{
bootstrap(df, m, strata = TRUE)
}
\arguments{
  \item{df}{is a data frame with the complete data set and
  class in the first column.} \item{m}{number of bootstrap
  replicates} \item{strata}{identifies if the bootstrap
  samples are stratified by class}
}
\value{
grouped data frame object with m bootstrap samples
}
\description{
Draws bootstrap samples with strata option.
}
\examples{
iris.b <- bootstrap(iris[, 5:1], 30, strata = FALSE)
index <- lapply(attributes(iris.b)$indices, function(x) x + 1)
}

