% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/bootstrap_pp.R
\name{bootstrap_pp}
\alias{bootstrap_pp}
\title{List of PP.Tree objects for bootstrap samples.}
\usage{
bootstrap_pp(data, scale = TRUE, size.p = 0.9, training, strata = TRUE,
  ntree, index = "LDA", ...)
}
\arguments{
\item{data}{is a data frame with the complete data set. Class factor in the first column}

\item{scale}{is to standardize the data set, by default it is TRUE.}

\item{training}{indices of the training data set}

\item{ntree}{number of bootstrap samples or individual classifiers (trees).}

\item{index}{PP index to run the PP.Tree, options LDA, Lp or PDA, by default it is LDA.}
}
\value{
list with the PP.Tree output for all the bootstraps samples and the index for the initial
  training set.
}
\description{
List of PP.Tree objects for bootstrap samples.
}
\examples{
training<-train_fn(iris[,5],.9)
data1<-iris[,5:1]
output<-bootstrap_pp(data1,scale=TRUE,size.p=.9,training,strata=TRUE,ntree=50,index="LDA")
}

