\name{ClassPP}
\alias{ClassPP}
\title{Projection Pursuit for Supervised Classification}
\description{
This package is for projection pursuit method for supervised classification.
}
 
\seealso{
PPindex.class \ PPindex.LDA \ PPindex.Lp \ PPindex.PDA

PP.optimize.anneal

PP.optimize.plot

PP.Tree

PP.classify
}
\references{
  Lee, E., Cook, D., and Klinke, S.(2002)
 \emph{Projection Pursuit indices for supervised classification}

}

\author{Eun-kyung Lee}

\keyword{multivariate}
