% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/crab.R
\docType{data}
\name{crab}
\alias{crab}
\title{Crab information of two crab species}
\format{A data frame with 200 rows and 6 variables}
\usage{
data(crab)
}
\description{
A data set containing 200 observations from two species of
crab (blue and orange). For each crab specie (50 in each
one) there are 50 males and 50 females.
}
\details{
\itemize{ \item{Type}{ class variable has 4 classes with
the combinations of specie and sex (BlueMale, BlueFemale,
OrangeMale and OrangeFemale)}. \item{FL}{the size of the
frontal lobe} \item{RW}{ rear width} \item{CL}{is carapace
length} \item{CW}{width} \item{BD}{body depth} }
}
\keyword{datasets}

