% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/fishcatch.R
\docType{data}
\name{fishcatch}
\alias{fishcatch}
\title{The Fish Catch data set contains measurements on 159 fish caught in the lake Laengelmavesi, Finland.}
\format{A data frame with 159 rows and 7 variables}
\usage{
data(fishcatch)
}
\description{
\itemize{ \item{Type}{has 7 fish classes, with 35 cases
of Bream, 11 cases of Parkki, 56 ases of Perch 17 cases
of Pike, 20 cases of Roach, 14 cases of Smelt and 6 cases
of Whitewish.} \item{weight}{Weight of the fish, in grams}
\item{length1}{Length from the nose to the beginning of
the tail, in cm} \item{length2}{Length from the nose to the
notch of the tail, in cm} \item{length3}{Length from the
nose to the end of the tail, in cm} \item{height}{Maximal
height as \% of Length3} \item{width}{Maximal width as \% of
Length3} }
}
\keyword{datasets}

