% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/forest_ppred.R
\name{forest_ppred}
\alias{forest_ppred}
\title{Vector with predicted values from a PPforest.}
\usage{
forest_ppred(newdata, output.tree, ...)
}
\arguments{
  \item{newdata}{are the complete data without the class
  variable we want to predict} \item{output.tree}{trees
  classifiers from trees_pp function or PPforest object}
  \item{...}{arguments to be passed to methods}
}
\value{
predicted values form PPforest
}
\description{
Vector with predicted values from a PPforest.
}
\examples{
data(iris)
iris.sc <- data.frame(Class = iris[, 5], scale(iris[, 1:4]))
training <- train_fn(class = iris[, 5], size.p = 2/3)
iris.b <- bootstrap( iris.sc[training$id, ], 500)
output <- trees_pp(iris.b, size.p = 0.9, PPmethod ="LDA")
pr <- forest_ppred( iris.sc[-training$id, 2:5] , output)
}

