% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glass.R
\docType{data}
\name{glass}
\alias{glass}
\title{Glass data set}
\format{A data frame with 214 rows and 10 variables}
\usage{
data(glass)
}
\description{
Contains measurements 214 observations of 6 types of glass; defined in terms of their oxide content.
\itemize{
\item{Type}{ has 6 types of glasses}
\item{X1}{ refractive index} 
\item{X2}{ Sodium (unit measurement: weight percent in corresponding oxide).} 
\item{X3}{ Magnesium}
\item{X4}{ Aluminum}
\item{X5}{ Silicon}
\item{X6}{ Potassium}
\item{X7}{ Calcium}
\item{X8}{ Barium}
\item{X9}{ Iron}
}
}
\keyword{datasets}
