% Generated by roxygen2 (4.0.2): do not edit by hand
\name{LDA.Tree}
\alias{LDA.Tree}
\title{Finding PP tree structure using LDA index}
\usage{
LDA.Tree( i.class, i.data, weight = TRUE, ...)
}
\arguments{
\item{i.class}{class information}

\item{i.data}{A training data  without class information}

\item{weight}{weight flag using in LDA index}
}
\value{
Tree.Struct Tree structure

Alpha.Keep 1D projection of each split

C.Keep spliting rule for each split
}
\description{
Find tree structure using linear discriminant in each split.
}
\examples{
data(iris)
n <- nrow(iris)
tot <- c(1:n)
n.train <- round(n*0.9)
train <- sample(tot,n.train)
test <- tot[-train]
}
\references{
Lee, E., Cook, D., Klinke, S.,  Lumley, T.(2005)
Projection Pursuit for Exploratory Supervised Classification,
Journal of Computational and Graphical Statistics, 14(4):831-846.
}
\seealso{
{\code{\link{PPindex.class}}, \code{\link{PP.optimize}}}
}
\keyword{tree}

