% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/leukemia.R
\docType{data}
\name{leukemia}
\alias{leukemia}
\title{Leukemia gene expession data set}
\format{A data frame with 72 rows and 41 variables}
\usage{
data(leukemia)
}
\description{
A data set containing 72 observations from 3 leukemia types
classes.
}
\details{
\itemize{ \item{Type}{ has 3 classes with 38 cases of B-cell
ALL, 25 cases of AML and 9 cases of T-cell ALL.} \item{Gene1
to Gen 40}{gene expression} }
}
\keyword{datasets}

