% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lymphoma.R
\docType{data}
\name{lymphoma}
\alias{lymphoma}
\title{Lymphoma data set}
\format{A data frame with 80 rows and 51 variables}
\usage{
data(lymphoma)
}
\description{
Gene expression in the three most prevalent adult lymphoid malignancies: B-cell chronic lymphocytic leukemia (B-CLL), follicular lymphoma (FL), and diffuse large B-cell lym- phoma (DLBCL). Gene expression levels were measured using a specialized cDNA microarray, the Lymphochip, containing genes that are preferentially expressed in lymphoid cells or that are of known immunologic or oncologic importance. This data set contain 80 observations from 3 lymphoma types.
\itemize{
  \item{Type}{ Class variable has 3 classes with 29 cases of B-cell ALL (B-CLL), 42 cases of diffuse large B-cell lymphoma (DLBCL) and 9 cases of follicular lymphoma (FL)}.
  \item{Gene1 to Gen 50}{gene expression}
  }
}
\keyword{datasets}

