% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/NCI60.R
\docType{data}
\name{NCI60}
\alias{NCI60}
\title{NCI60 gene expession data set}
\format{A data frame with 61 rows and 31 variables}
\usage{
data(NCI60)
}
\description{
A data set containing 61 observations from 8 different
tissue types.
}
\details{
\itemize{ \item{Type}{ has 8 different tissue types, 9
cases of breast, 5 cases of central nervous system (CNS), 7
cases pf colon, 8 cases of leukemia, 8 cases of melanoma, 9
cases of non-small-cell lung carcinoma (NSCLC), 6 cases of
ovarian and 9 cases of renal.} \item{Gene1 to Gen 30}{gene
expression} }
}
\keyword{datasets}

