% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/node_data.R
\name{node_data}
\alias{node_data}
\title{Data structure with the  projected and boundary by node and class.}
\usage{
node_data(ppf, tr, Rule = 1)
}
\arguments{
\item{ppf}{is a PPforest object}

\item{tr}{numerical value to  identify a tree}

\item{Rule}{split rule 1:mean of two group means, 2:weighted mean, 3: mean of max(left group) and min(right group), 4: weighted mean of max(left group) and min(right group)}
}
\value{
Data frame with projected data for each class and node id and the boundaries
}
\description{
Data structure with the  projected and boundary by node and class.
}
\examples{
#crab data set with all the observations used as training

pprf.crab <- PPforest(data = crab, std = 'min-max', y = 'Type',
 size.tr = 1, m = 200, size.p = .5, PPmethod = 'LDA')
node_data(ppf = pprf.crab, tr = 1) 

}
