% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olive.R
\docType{data}
\name{olive}
\alias{olive}
\title{The olive data set}
\format{A data frame contains 573 observations and 10 variables}
\usage{
data(olive)
}
\description{
contains  572 observations and 10 variables
\itemize{
\item{Region}{Three super-classes of Italy: North, South and the island of Sardinia }
\item{area}{ Nine collection areas: three from North, four from South and 2 from Sardinia} 
\item{palmitic}{ fatty acids percent x 100}
\item{palmitoleic}{ fatty acids percent x 100 }
\item{stearic}{fatty acids percent x 100}
\item{oleic}{fatty acids percent x 100 }
\item{linoleic}{fatty acids percent x 100}
\item{linolenic}{fatty acids percent x 100  }
\item{arachidic}{ fatty acids percent x 100}
\item{eicosenoic}{ fatty acids percent x 100}
}
}
\keyword{datasets}
