% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_dat.R
\name{parallel_dat}
\alias{parallel_dat}
\title{Parallel plot for the data set colored by class, the data are standarized in the function}
\usage{
parallel_dat(data, class)
}
\arguments{
\item{data}{Data frame with the complete data set.}

\item{class}{A character with the name of the class variable.}
}
\value{
A parallel coordinate plot with the data selt collored by class
}
\description{
Parallel plot for the data set colored by class, the data are standarized in the function
}
\examples{
#crab data set with all the observations used as training
pprf.crab <- PPforest2(data =crab, class = "Type",
 size.tr = 1, m = 200, size.p = .5, PPmethod = 'LDA', strata = TRUE)
parallel_dat(crab,"Type")
}

