% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/parkinson.R
\docType{data}
\name{parkinson}
\alias{parkinson}
\title{parkinson data set}
\format{A data frame with 195 rows and 23 variables}
\usage{
data(parkinson)
}
\description{
A data set containing 195 observations from 2 parkinson
types.
}
\details{
\itemize{ \item{Type}{ Class variable has 2 classes, there
are 48 cases of healthy people and 147 cases with Parkinson.
The feature variables are biomedical voice measures}.
\item{X1 to X22}{Check vbles} }
}
\keyword{datasets}

