% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/PDA.Tree.R
\name{PDA.Tree}
\alias{PDA.Tree}
\title{Find PP tree structure using PDA index}
\usage{
PDA.Tree(i.class, i.data, weight = TRUE, lambda=1, ...)
}
\arguments{
\item{i.class}{class information}

\item{i.data}{A training data  without class information}

\item{weight}{weight flag using in LDA index}

\item{lambda}{a parameter for PDA index}
}
\value{
Tree.Struct Tree structure

Alpha.Keep 1D projection of each split

C.Keep spliting rule for each split
}
\description{
Find tree structure using projection pursuit in each split.
}
\examples{
data(iris)
n <- nrow(iris)
tot <- c(1:n)
n.train <- round(n*0.9)
train <- sample(tot,n.train)
test <- tot[-train]
Tree.result <- PDA.Tree(iris[train,5],iris[train,1:4])
Tree.result
}
\references{
Lee, E.-K., Cook, D. (2009) A Projection Pursuit Index for Large p Small n Data,
Statistics and Computing,
\url{http://www.springerlink.com/content/g47n0n342761838m/#?p=d2ff5a7b69eb45ef8abf7ef3aba69557&pi=3}.
}
\seealso{
{\code{\link{PPindex.class}}, \code{\link{PP.optimize}}}
}
\keyword{tree}

