% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_importance.R
\name{permute_importance}
\alias{permute_importance}
\title{Obtain the importance variable measure based in permutation}
\usage{
permute_importance(ppf, cl.tree = TRUE, pl = TRUE)
}
\arguments{
\item{ppf}{is a PPforest object}

\item{cl.tree}{when it is TRUE compute the importance variable base on permuted oob observations for each variable when it is FALSE compute an importance for all the forest with permuted values for each variable instead}

\item{pl}{plot is TRUE by default}
}
\value{
importance variable plot
}
\description{
Obtain the importance variable measure based in permutation
}
\examples{
pprf.crab<- PPforest(data = crab, class = "Type",
 size.tr = 1, m = 100, size.p = .4, PPmethod = 'LDA', strata = TRUE)
permute_importance(ppf = pprf.crab, cl.tree = FALSE, pl = TRUE) 
}

