% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_imortance2.R
\name{permute_importance2}
\alias{permute_importance2}
\title{Obtain the importance variable measure based in permutation}
\usage{
permute_importance2(ppf)
}
\arguments{
\item{ppf}{is a PPforest object}
}
\value{
A data frame with permuted importance measure
}
\description{
Obtain the importance variable measure based in permutation
}
\examples{
pprf.crab <- PPforest2(data = crab, class = "Type",
 size.tr = 1, m = 100, size.p = .4, PPmethod = 'LDA', strata = TRUE)
permute_importance2(ppf = pprf.crab) 

}

