% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/pp.ternary.plot.R
\name{pp.ternary.plot}
\alias{pp.ternary.plot}
\title{Ternary plot}
\usage{
pp.ternary.plot(x, y, z, pprf, data)
}
\arguments{
\item{x}{is a numeric value which identifies the selected class to plot in the left corner.}

\item{y}{is a numeric value which identifies the selected class to plot in the right corner.}

\item{z}{is a numeric value which identifies the selected class to plot in the top corner.}

\item{pprf}{a PPforest object}

\item{data}{a data frame with the complete data set}
}
\description{
\code{pp.ternary.plot} implements a ternary plot .
}
\examples{
#crab data set without test
pprf.crab2 <- PPforest(y = crab[, 1], x = crab[, -1], size.tr = 1,
m = 100, size.p = .8, PPmethod = 'LDA', strata = TRUE)
pp.ternary.plot(1, 2, 3, pprf.crab2, crab)
}

