% Generated by roxygen2 (4.0.2): do not edit by hand
\name{PP.Tree}
\alias{PP.Tree}
\title{Find PP tree structure using PDA, Lp or PDA indexes}
\usage{
PP.Tree(PPmethod, i.class, i.data, weight = TRUE, r = NULL,
lambda = NULL, cooling = 0.999, temp = 1, energy = 0.01,  ...)
}
\arguments{
\item{i.class}{class information}

\item{i.data}{A training data  without class information}

\item{weight}{weight flag using in LDA index}

\item{r}{a parameter for L_r index}

\item{lambda}{a parameter for PDA index}

\item{cooling}{parameter for optimization}

\item{temp}{inital temperature for optimization}

\item{energy}{parameter for simulated annealing optimization}
}
\value{
Tree.Struct Tree structure

Alpha.Keep 1D projection of each split

C.Keep spliting rule for each split
}
\description{
Find tree structure using projection pursuit in each split.
}
\examples{
data(iris)
n <- nrow(iris)
tot <- c(1:n)
n.train <- round(n*0.9)
train <- sample(tot,n.train)
test <- tot[-train]
Tree.result <- PP.Tree("LDA",iris[train,5],iris[train,1:4])
Tree.result
}
\references{
Lee, E., Cook, D., and Klinke, S.(2002) Projection Pursuit indices for supervised
classification
}
\seealso{
{\code{\link{PPindex.class}}, \code{\link{PP.optimize}}}
}
\keyword{tree}

