% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPclassify.R
\name{PPclassify}
\alias{PPclassify}
\title{Predict class for the test set and calculate prediction error after finding the PPtree structure, .}
\usage{
PPclassify( Tree.result, test.data = NULL, Rule = 1, true.class = NULL)
}
\arguments{
\item{Tree.result}{the result from PPtree_split}

\item{test.data}{the test dataset}

\item{Rule}{split rule 1:mean of two group means, 2:weighted mean, 3: mean of max(left group) and min(right group), 4: weighted mean of max(left group) and min(right group)}

\item{true.class}{true class of test dataset if available}
}
\value{
predict.class predicted class

predict.error prediction error
}
\description{
Predict class for the test set and calculate prediction error after finding the PPtree structure, .
}
\examples{
#crab data set
set.seed(143)
idx <-sample(1:200, 150)
Tree.crab <- PPtree_split('Type~.', data = crab[idx,], PPmethod = 'LDA', size.p = 1)
Tree.crab
PPclassify(Tree.crab, test.data = crab[-idx, 2:6], Rule = 1,true.class = crab[-idx, 1])


}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013) 
PPtree: Projection pursuit classification tree, 
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}
