% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPclassify2.R
\name{PPclassify2}
\alias{PPclassify2}
\title{Predict class for the test set and calculate prediction error}
\usage{
PPclassify2( Tree.result, test.data, Rule, true.class)
}
\arguments{
\item{Tree.result}{the result of PP.Tree}

\item{test.data}{the test dataset}

\item{Rule}{split rule 1:mean of two group means, 2:weighted mean, 3: mean of max(left group) and min(right group), 4: weighted mean of max(left group) and min(right group)}

\item{true.class}{true class of test dataset if available}
}
\value{
predict.class predicted class

predict.error prediction error
}
\description{
After finding tree structure, predict class for the test set and calculate prediction error.
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013) 
PPtree: Projection pursuit classification tree, 
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}

