% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppf_bootstrap.R
\name{ppf_bootstrap}
\alias{ppf_bootstrap}
\title{Draws bootstrap samples with strata option.}
\usage{
ppf_bootstrap(data, class, m = 500, strata = TRUE)
}
\arguments{
\item{data}{Data frame with the complete data set.}

\item{class}{A character with the name of the class variable.}

\item{m}{is the number of bootstrap replicates we want to sample.}

\item{strata}{is TRUE if the bootstrap samples are stratified by class variable.}
}
\value{
data frame object with m bootstrap samples
}
\description{
Draws bootstrap samples with strata option.
}
\examples{
leukemia.b <- ppf_bootstrap(data = leukemia, class = "Type",  m = 200) 
index <- lapply(attributes(leukemia.b)$indices, function(x) x + 1)
}

