% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppf_global_imp.R
\name{ppf_global_imp}
\alias{ppf_global_imp}
\title{Global importance measure for a PPforest object}
\usage{
ppf_global_imp(data, class, ppf)
}
\arguments{
\item{data}{Data frame with the complete data set.}

\item{class}{A character with the name of the class variable.}

\item{ppf}{is a PPforest object}
}
\value{
Data frame with the global importance measure
}
\description{
Global importance measure for a PPforest object
}
\examples{
#crab data set with all the observations used as training

pprf.crab <- PPforest(data = crab, std = TRUE, class = 'Type',
 size.tr = 1, m = 200, size.p = .5, PPmethod = 'LDA', parallel = TRUE, cores = 2)
 
ppf_global_imp(data = crab, class = 'Type', pprf.crab) 

}
