% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppf_importance.R
\name{ppf_importance}
\alias{ppf_importance}
\title{Importance variable visualization}
\usage{
ppf_importance(data, class, ppf, global = TRUE, weight = TRUE, interactive)
}
\arguments{
\item{data}{Data frame with the complete data set.}

\item{class}{A character with the name of the class variable.}

\item{ppf}{is a PPforest object}

\item{global}{is TRUE if we want to see the global importance of the forest}

\item{weight}{is TRUE if we want to see a weighted mesure of the forest importance based on out of bag trees errors}

\item{interactive}{if is TRUE will use plotly to translate an static ggplot object}
}
\value{
A dotplot with a global measure of importance  variables in the PPforest.
}
\description{
Importance variable visualization
}
\examples{
#crab data set with all the observations used as training
pprf.crab <- PPforest(data = crab, class = "Type",
 size.tr = 1, m = 200, size.p = .5, PPmethod = 'LDA')
ppf_importance(data = crab, class = "Type", pprf.crab, global = TRUE,
 weight = FALSE, interactive = TRUE) 
}

