% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/print.PPforest.R
\name{print.PPforest}
\alias{print.PPforest}
\title{Print PPforest object}
\usage{
\method{print}{PPforest}(x, ...)
}
\arguments{
\item{x}{is a PPforest class object}

\item{...}{additional parameter}
}
\value{
printed results for PPforest object
}
\description{
Print PPforest object
}

