% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.PPforest2.R
\name{print.PPforest2}
\alias{print.PPforest2}
\title{Print PPforest object}
\usage{
\method{print}{PPforest2}(x, ...)
}
\arguments{
\item{x}{is a PPforest class object}

\item{...}{additional parameter}
}
\value{
printed results for PPforest object
}
\description{
Print PPforest object
}

