% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ternary_str.R
\name{ternary_str}
\alias{ternary_str}
\title{Data structure with the  projected and boundary by node and class.}
\usage{
ternary_str(ppf, id, sp, dx, dy)
}
\arguments{
\item{ppf}{is a PPforest object}

\item{id}{is a vector with the selected projection directions}

\item{sp}{is the simplex dimensions, if k is the number of classes sp = k - 1}

\item{dx}{first direction included in id}

\item{dy}{second direction included in id}
}
\value{
Data frame needed to visualize a ternary plot
}
\description{
Data structure with the  projected and boundary by node and class.
}
\examples{
#crab data set with all the observations used as training
pprf.crab <- PPforest(data = crab, std =TRUE, class = "Type",
 size.tr = 1, m = 200, size.p = .5, PPmethod = 'LDA')
ternary_str ( pprf.crab, id = c(1,2,3), sp = 3, dx = 1, dy = 2) 
}

