% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/train_fn.R
\name{train_fn}
\alias{train_fn}
\title{Obtain stratified sample of the data}
\usage{
train_fn(data, class, size.p = 0.7)
}
\arguments{
\item{data}{Data frame with the complete data set.}

\item{class}{A character with the name of the class variable.}

\item{size.p}{Sample size proportion in each class group.}
}
\value{
numeric vector  with the sample indexes for training data
}
\description{
Index id for training set, sample in each class with the same proportion.
}
\examples{
training.id <- train_fn(data = leukemia, class = "Type", size.p = 0.7)
training.id
}

