% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_ppred.R
\name{tree_ppred}
\alias{tree_ppred}
\title{Obtain predicted class for new data using PPforest}
\usage{
tree_ppred(xnew, output.tree, ...)
}
\arguments{
\item{xnew}{data frame with explicative variables used to get new predicted values.}

\item{output.tree}{trees classifiers from trees_pp function or PPforest object}

\item{...}{arguments to be passed to methods}
}
\value{
predicted values form PPforest
}
\description{
Vector with predicted values from a PPforest.
}
\examples{
training.id <- train_fn(y = leukemia[,1], size.p = 0.9)
leukemia.b <- ppf_bootstrap(y = leukemia[, 1], df = leukemia, m= 200) 
leukemia.trees <- trees_pp(data.b = leukemia.b, size.p = .9, PPmethod = 'PDA', lambda = .1)
pr <- tree_ppred( xnew = leukemia[-training.id$id, -1] , leukemia.trees)
pr
}

