% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_pp.R
\name{trees_pp}
\alias{trees_pp}
\title{Grow a PPtree_split for each bootstrap sample}
\usage{
trees_pp(data.b, size.p = 0.9, PPmethod = "LDA", lambda = 0.1, ...)
}
\arguments{
\item{data.b}{are the bootstrap samples from training set.}

\item{size.p}{proportion of random sample variables in each split.}

\item{PPmethod}{is the projection pursuit index to be optimized, options LDA or PDA, by default it is LDA.}

\item{lambda}{a parameter for PDA index}

\item{...}{arguments to be passed to methods}
}
\value{
data frame with trees_pp output for all the bootstraps samples.
}
\description{
For each bootstrap sample grow a projection persuit tree (PPtree object).
}
\examples{
#leukemia data set
leukemia.b <- ppf_bootstrap(data = leukemia, class = "Type", m = 70) 
leukemia.trees <- trees_pp(data.b = leukemia.b, size.p = .4, PPmethod = 'PDA', lambda = .1) 
str(leukemia.trees,max.level=1)
}

