% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_pred.R
\name{trees_pred}
\alias{trees_pred}
\title{Obtain predicted class for new data from baggtree function}
\usage{
trees_pred(object, xnew, parallel = FALSE, cores = 2, ...)
}
\arguments{
\item{object}{trees classifiers from trees_pp function or PPforest object}

\item{xnew}{data frame with explicative variables used to get new predicted values.}

\item{...}{arguments to be passed to methods}
}
\value{
predicted values form PPforest
}
\description{
Vector with predicted values from a baggtree function.
}
\examples{
crab.trees <- baggtree(data = crab, class = "Type", 
m =  200, PPmethod = 'LDA', lambda = .1, size.p = 0.4 )
 
pr <- trees_pred(  crab.trees,xnew = crab[, -1] )


pprf.crab <- PPforest(data = crab, class = "Type",
 std = TRUE, size.tr = 1, m = 100, size.p = .4, PPmethod = 'LDA' )
 
trees_pred(pprf.crab, xnew = pprf.crab$train[, -1] )

}
