% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/importance.R
\name{varimp_pp}
\alias{varimp_pp}
\title{Data frame with the 1D projection of each split weighted by 1- oob error rate for each tree and plots of importance variable for each split and jointly.}
\usage{
varimp_pp(data, boot, bagg)
}
\arguments{
\item{data}{is a data frame with the complete data set. Class factor in the first column}

\item{boot}{object from bootstrap_pp (list of PP.Tree objets for bootstrap samples)}

\item{bagg}{object form bagging_pp (list with bagging PP.tree error and bagging classification)}
}
\value{
a data frame with 1D projection of each split (Alpha.Keep) and a boxplot with the importance of variable for each split.
}
\description{
Data frame with the 1D projection of each split weighted by 1- oob error rate for each tree and plots of importance variable for each split and jointly.
}
\examples{
data1<-iris[,5:1]
training<-train_fn(iris[,5],.9)
output<-bootstrap_pp(data1,scale=TRUE,training,ntree=50,index="LDA")
b.pp <- bagging_pp(data1,scale=TRUE, strata=TRUE,output,training)
varimp_pp(data1,output,b.pp)
}

