% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/wine.R
\docType{data}
\name{wine}
\alias{wine}
\title{wine data set}
\format{A data frame with 178 rows and 14 variables}
\usage{
data(wine)
}
\description{
A data set containing 178 observations from 3 wine grown cultivares in Italy.
}
\details{
\itemize{
  \item{Type}{ Class variable has 3 classes that are 3 different wine grown cultivares in Italy. }
  \item{X1 to X13}{Check vbles}
}
}
\keyword{datasets}

