% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/bagging_pp.R
\name{bagging_pp}
\alias{bagging_pp}
\title{List with bagging PP.tree error and bagging classification.}
\usage{
bagging_pp(data, scale = TRUE, strata = TRUE, boot, training, ...)
}
\arguments{
  \item{data}{is a data frame with the data to classify.
  Class factor in the first column}

  \item{scale}{is to standardize the data set, by default
  it is TRUE.}

  \item{strata}{is to stratify the bootstrap samples, by
  default is TRUE.}

  \item{boot}{object from bootstrap_pp (list of PP.Tree
  objets for bootstrap samples)}

  \item{training}{indices of the training data set.}
}
\value{
list with the error, predicted classes and confusion
matrix.
}
\description{
List with bagging PP.tree error and bagging classification.
}
\examples{
data1<-iris[,5:1]
training<-train_fn(iris[,5],.9)
output<-bootstrap_pp(data1,scale=TRUE,size.p=.9,training,strata=TRUE,ntree=50,index="LDA")
b.pp <- bagging_pp(data1,scale=TRUE, strata=TRUE,output,training)
}

