% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/bootstrap_pp.R
\name{bootstrap_pp}
\alias{bootstrap_pp}
\title{List of PP.Tree objects for bootstrap samples.}
\usage{
bootstrap_pp(data, training, ntree, index = "LDA", ...)
}
\arguments{
\item{data}{is a data frame with the complete data set. Class factor in the first column}

\item{training}{indices of the training data set}

\item{ntree}{number of bootstrap samples or individual classifiers (trees).}

\item{index}{PP index to run the PP.Tree, options LDA, Lp or PDA.}

\item{size.p}{proportion of sample in each class group.}
}
\value{
list with the PP.Tree output for all the bootstraps samples and the index for the initial
  training set.
}
\description{
List of PP.Tree objects for bootstrap samples.
}
\examples{
data1<-iris[,5:1]
output<-bootstrap_pp(data1,training,ntree=50,index="LDA")
output
}

