% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crab.R
\docType{data}
\name{crab}
\alias{crab}
\title{Astralian crabs}
\format{
Measurements on rock crabs of the genus Leptograpsus. The data set contains 200 observations
 from two species of crab (blue and orange), there are 50 specimens of each sex of each species,
  collected on site at Fremantle, Western Australia. 
\describe{
  \item{Type}{ is the class variable and has 4 classes with the combinations of species and sex (BlueMale, BlueFemale, OrangeMale and OrangeFemale).} 
  \item{FL}{the size of the frontal lobe length, in mm.}
  \item{RW}{rear width, in mm}
  \item{CL}{length of midline of the carapace, in mm.}
  \item{CW}{maximum width of carapace, in mm.}
  \item{BD}{depth of the body; for females, measured after displacement of the abdomen, in mm.}
}

A data frame with 200 rows and 6 variables
}
\source{
Campbell, N. A. & Mahon, R. J. (1974), A Multivariate Study of Variation in Two Species of Rock Crab of genus Leptograpsus, Australian Journal of Zoology 22(3), 417 - 425.
}
\usage{
data(crab)
}
\description{
Astralian crabs
}
\keyword{datasets}
