% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishcatch.R
\docType{data}
\name{fishcatch}
\alias{fishcatch}
\title{Fish catch data set}
\format{
A data frame with 159 rows and 7 variables
}
\source{
url{http://www.amstat.org/publications/jse/jse_data_archive.htm}
}
\usage{
data(fishcatch)
}
\description{
There are 159 fishes of 7 species are caught and measured. Altogether there are
  7 variables.  All the fishes are caught from the same lake(Laengelmavesi) near Tampere in Finland.
\itemize{
\item{Type}{ has 7 fish classes, with 35 cases of Bream, 11 cases of Parkki, 56 cases of Perch 17 cases of Pike, 20 cases of Roach, 14 cases of Smelt and 6 cases of Whitewish.} 
\item{weight}{ Weight of the fish (in grams)}
\item{length1}{ Length from the nose to the beginning of the tail (in cm)}
\item{length2}{ Length from the nose to the notch of the tail (in cm)}
\item{length3}{ Length from the nose to the end of the tail (in cm)}
\item{height}{ Maximal height as \% of Length3}
\item{width}{ Maximal width as \% of Length3}
}
}
\keyword{datasets}
