% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Data frame with the 1D projection of each split unweighted and weighted by 1-(oob error rate) for each tree and plots of importance variable for each split.}
\usage{
importance(data, ppforest, global = TRUE, weight = TRUE)
}
\arguments{
  \item{data}{is a data frame with the complete data set. Class factor in the first column} 
  \item{ppforest}{is an object from PPforest} 
  \item{global}{is TRUE if we wan to see the global importance of the forest}
  \item{weight}{is TRUE if we want to see a weighted mesure of the forest importance}
}
\value{
a data frame with 1D projection of each split (Alpha.Keep)
and a boxplot with the importance of variable for each
split.
}
\description{
Data frame with the 1D projection of each split unweighted
and weighted by 1-(oob error rate) for each tree and plots
of importance variable for each split.
}
\examples{
tr.index <- train_fn(iris[, 5], 2/3)
te.index <- as.vector(1:length(iris[, 5]))[!(1:length(iris[, 5]) \%in\% (sort(tr.index$id)))]
train <- iris[sort(tr.index$id), 5:1]
test <- iris[-tr.index$id, 5:1 ]
ppforest <- PPforest( train = train, testap = TRUE, test = test,
m = 500, size.p = .9, PPmethod = 'LDA', strata = TRUE)
importance(train, ppforest)
}

