% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/lymphoma.R
\docType{data}
\name{lymphoma}
\alias{lymphoma}
\title{Lymphoma gene expession data set}
\format{A data frame with 80 rows and 51 variables}
\usage{
data(lymphoma)
}
\description{
A data set containing 80 observations from 3 lymphoma types.
}
\details{
\itemize{ \item{Type}{ Class variable has 3 classes with 29
cases of B-cell ALL (B-CLL), 42 cases of diffuse large B-
cell lymphoma (DLBCL) and 9 cases of follicular lymphoma
(FL)}. \item{Gene1 to Gen 50}{gene expression} }
}
\keyword{datasets}

