% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NCI60.R
\docType{data}
\name{NCI60}
\alias{NCI60}
\title{NCI60 data set}
\format{
A data frame with 61 rows and 31 variables
}
\source{
Dudoit, S., Fridlyand, J. and Speed, T. P. (2002). Comparison of Discrimination Methods for the Classification of Tumors Using Gene Expression Data. Journal of the American statistical Association 97 77-87.
}
\usage{
data(NCI60)
}
\description{
cDNA microarrays were used to examine the variation in gene expression among the 60 cell lines.  The cell lines are derived from tumors with different sites of origin. This data set contain 61 observations and 30 feature variables from 8 different tissue types.
\itemize{
  \item{Type}{ has 8 different tissue types, 9 cases of breast, 5 cases of central nervous system (CNS), 7 cases pf colon, 8 cases of leukemia, 8 cases of melanoma, 9 cases of  non-small-cell lung carcinoma (NSCLC), 6 cases of ovarian and 9 cases of renal.} 
  \item{Gene1 to Gen 30}{ gene expression information}
}
}
\keyword{datasets}
