% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_dat.R
\name{parallel_dat}
\alias{parallel_dat}
\title{Parallel plot for the data set colored by class}
\usage{
parallel_dat(data, class)
}
\arguments{
\item{data}{Data frame with the complete data set.}

\item{class}{A character with the name of the class variable.}
}
\value{
A parallel coordinate plot with the data selt collored by class
}
\description{
Parallel plot for the data set colored by class
}
\examples{
#leukemia data set with all the observations used as training
pprf.leukemia <- PPforest(data = leukemia, class = "Type",
 size.tr = 1, m = 70, size.p = .4, PPmethod = 'PDA', strata = TRUE)
parallel_dat(leukemia,"Type")
}

