% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permute_importance.R
\name{permute_importance}
\alias{permute_importance}
\title{Obtain the permuted importance variable measure}
\usage{
permute_importance(ppf)
}
\arguments{
\item{ppf}{is a PPforest object}
}
\value{
A data frame with permuted importance measures, imp is the permuted importance measure defined in Brieman paper,
imp2 is the permuted importance measure defined in randomForest package, the standard deviation (sd.im and sd.imp2) for each measure is computed and the 
also the standardized mesure.
}
\description{
Obtain the permuted importance variable measure
}
\examples{
pprf.crab <- PPforest(data = crab, class = 'Type',
std = TRUE, size.tr = 1, m = 100, size.p = .4, PPmethod = 'LDA', parallel = TRUE, core = 2)
permute_importance(ppf = pprf.crab) 

}
