% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PPclassify.R
\name{PPclassify}
\alias{PPclassify}
\title{Predict class for the test set and calculate prediction error}
\usage{
PPclassify( Tree.result, test.data = NULL, Rule = 1, true.class = NULL)
}
\arguments{
\item{Tree.result}{the result of PP.Tree}

\item{test.data}{the test dataset}

\item{Rule}{split rule 1:mean of two group means, 2:weighted mean, 3: mean of max(left group) and min(right group), 4: weighted mean of max(left group) and min(right group)}

\item{true.class}{true class of test dataset if available}
}
\value{
predict.class predicted class

predict.error prediction error
}
\description{
After finding tree structure, predict class for the test set and calculate prediction error.
}
\examples{
#crab data set
Tree.crab <- PPtree_split('Type~.', data = crab, PPmethod = 'LDA', size.p = 0.5)
Tree.crab

PPclassify(Tree.crab)
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK(2013) 
PPtree: Projection pursuit classification tree, 
Electronic Journal of Statistics, 7:1369-1386.
}
\keyword{tree}
