% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ppf_bootstrap.R
\name{ppf_bootstrap}
\alias{ppf_bootstrap}
\title{Draws bootstrap samples with strata option.}
\usage{
ppf_bootstrap(y, df, m = 500, strata = TRUE)
}
\arguments{
\item{y}{is a vector with the class variable.}

\item{df}{is a data frame with class and explicative variables.}

\item{m}{is the number of bootstrap replicates we want to sample.}

\item{strata}{is TRUE if the bootstrap samples are stratified by class variable.}
}
\value{
data frame object with m bootstrap samples
}
\description{
Draws bootstrap samples with strata option.
}
\examples{
leukemia.b <- ppf_bootstrap(y = leukemia[, 1], df = leukemia, m= 200)
index <- lapply(attributes(leukemia.b)$indices, function(x) x + 1)
}

