% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ppf_importance.R
\name{ppf_importance}
\alias{ppf_importance}
\title{Importance variable visualization}
\usage{
ppf_importance(y, x, ppforest, global = TRUE, weight = TRUE)
}
\arguments{
\item{y}{is a vector with the class variable.}

\item{x}{is a data frame with explicative variables,  should be standardized.}

\item{ppforest}{is a PPforest object}

\item{global}{is TRUE if we want to see the global importance of the forest}

\item{weight}{is TRUE if we want to see a weighted mesure of the forest importance based on out of bag trees errors}
}
\value{
A dotplot with a global measure of importance  variables in the PPforest.
}
\description{
Importance variable visualization
}
\examples{
#leukemia data set with 2/3 observations used as training
pprf.leukemia <- PPforest(y = leukemia[, 1], x = leukemia[, -1],
 size.tr = 2/3, m = 500, size.p = .5, PPmethod = 'PDA', strata = TRUE)
ppf_importance(y = leukemia[, 1], x = leukemia[, -1], pprf.leukemia,
global = TRUE, weight = TRUE)
}

