% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ppf_importance.R
\name{ppf_importance}
\alias{ppf_importance}
\title{Importance variable visualization}
\usage{
ppf_importance(data, ppforest, global = TRUE, weight = TRUE)
}
\arguments{
\item{data}{is a data frame with the complete data set. Class factor in the first column and the data should be standardize}

\item{ppforest}{is a PPforest object}

\item{global}{is TRUE if we want to see the global importance of the forest}

\item{weight}{is TRUE if we want to see a weighted mesure of the forest importance based on out of bag trees errors}
}
\value{
A dotplot with a global measure of importance  variables in the PPforest.
}
\description{
Importance variable visualization
}
\examples{
iris.s<- data.frame(iris[, 5], scale(iris[, 1:4]))
names(iris.s) <- colnames(iris[, 5:1])
ppforest <- PPforest( data = iris.s, size.tr = 2/3,
m = 500, size.p = .9, PPmethod = 'LDA', strata = TRUE)
ppf_importance(iris[,5:1], ppforest, global = TRUE, weight = TRUE)
}

