% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppf_importance.R
\name{ppf_importance}
\alias{ppf_importance}
\title{Importance variable visualization}
\usage{
ppf_importance(data, class, ppf, global = TRUE, weight = TRUE)
}
\arguments{
\item{data}{Data frame with the complete data set.}

\item{class}{A character with the name of the class variable.}

\item{ppf}{is a PPforest object}

\item{global}{is TRUE if we want to see the global importance of the forest}

\item{weight}{is TRUE if we want to see a weighted mesure of the forest importance based on out of bag trees errors}
}
\value{
A dotplot with a global measure of importance  variables in the PPforest.
}
\description{
Importance variable visualization
}
\examples{
#leukemia data set with all the observations used as training
pprf.leukemia <- PPforest(data = leukemia, class = "Type",
 size.tr = 1, m = 70, size.p = .4, PPmethod = 'PDA', strata = TRUE)
ppf_importance(data = leukemia, class = "Type", pprf.leukemia, global = TRUE, weight = TRUE) 
}

