% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppf_importance2.R
\name{ppf_importance2}
\alias{ppf_importance2}
\title{Importance variable visualization for each tree in the forest}
\usage{
ppf_importance2(data, class, ppf, tree.id = 1)
}
\arguments{
\item{data}{Data frame with the complete data set.}

\item{class}{A character with the name of the class variable.}

\item{ppf}{is a PPforest object}

\item{tree.id}{id number of the tree we want to see the specific importance measure will be in red}
}
\value{
A side-by-side jittered dotplot with the absolute importance measure for each tree in the forest
}
\description{
Importance variable visualization for each tree in the forest
}
\examples{
#leukemia data set with all the observations used as training
pprf.leukemia <- PPforest(data = leukemia, class = "Type",
 size.tr = 1, m = 70, size.p = .4, PPmethod = 'PDA', strata = TRUE)
ppf_importance2(leukemia,"Type",pprf.leukemia, tree.id = 1)
}

