% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/ppf_oob_error.R
\name{ppf_oob_error}
\alias{ppf_oob_error}
\title{OOB error visualization}
\usage{
ppf_oob_error(ppf, nsplit)
}
\arguments{
\item{ppf}{a PPforest object}

\item{nsplit}{number of points where cumulative oob error rate is computed}
}
\value{
a plot with the cumulative oob error rate
}
\description{
\code{ppf_oob_error} Plot the cummulative oob error as a function of number of trees
}
\examples{
pprf.leukemia <- PPforest(y = leukemia[, 1], x = leukemia[, -1],
size.tr = 2/3, m = 500, size.p = .5, PPmethod = 'PDA', strata = TRUE)
ppf_oob_error(ppf = pprf.leukemia, nsplit = 50)
}

