% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ppf_oob_error.R
\name{ppf_oob_error}
\alias{ppf_oob_error}
\title{OOB error visualization}
\usage{
ppf_oob_error(ppfo, nsplit)
}
\arguments{
\item{ppfo}{a PPforest object}

\item{nsplit}{number of points where cumulative oob error rate is computed}
}
\value{
a plot with the cumulative oob error rate
}
\description{
\code{ppf_oob_error} Plot the cummulative oob error as a function of number of trees
}
\examples{
pprf.crab <- PPforest(data = crab, size.tr = 1, m = 500, size.p = .5,
PPmethod = 'LDA',  strata = TRUE)
ppf_oob_error(pprf.crab, nsplit = 20)
}

