% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppf_oob_error.R
\name{ppf_oob_error}
\alias{ppf_oob_error}
\title{OOB error visualization}
\usage{
ppf_oob_error(ppf, nsplit1, nsplit2)
}
\arguments{
\item{ppf}{a PPforest object}

\item{nsplit1}{number, increment of the sequence where cumulative oob error rate is computed in the  1/3 trees}

\item{nsplit2}{number, increment of the sequence where cumulative oob error rate is computed in the  2/3 trees}
}
\value{
a plot with the cumulative oob error rate
}
\description{
\code{ppf_oob_error} Plot the cummulative oob error as a function of number of trees
}
\examples{
pprf.leukemia <- PPforest(data = leukemia, class = "Type", 
size.tr = 1, m = 70, size.p = .4, PPmethod = 'PDA', strata = TRUE)
ppf_oob_error(ppf = pprf.leukemia, nsplit1 = 5, nsplit2 = 15)
}

