% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/PPforest.R
\name{PPforest}
\alias{PPforest}
\title{Run a Projection Pursuit forest}
\usage{
PPforest(train, testap = TRUE, test, m, PPmethod, size.p, strata = TRUE,
  lambda = 0.1)
}
\arguments{
  \item{train}{is a data frame with the training data with class in the first column} 
  \item{testap}{is TRUE if we want to predict a new data set} 
  \item{test}{is a data frame it the new test data to predict}
  \item{m}{number of bootstrap replicate} 
  \item{PPmethod}{to run the PPtree_plit function, options LDA or PDA, by default it is LDA.}                 \item{size.p}{proportion of random sample variables in each split.} 
  \item{strata}{identify if the bootrap samples are stratify by class} 
  \item{lambda}{a parameter for PDA index}
}
\value{
An object of class \code{PPforest} with components
\item{prediction.training}{predicted values for training data set} 
\item{training.error}{error of the training data set}
\item{prediction.test}{predicted values for the test data set} 
\item{error.test}{error of the test data set}
\item{oob.error.forest}{out of bag error in the forest}
\item{oob.error.tree}{out of bag error for each tree}
\item{boot.samp}{bootrap samples} 
\item{output.trees}{output from a trees_pp} 
\item{proximity}{Proximity matrix} 
\item{vote.matrix}{Marix with votes for each class on each observation} 
\item{n.tree}{number of trees in \code{PPforest}} 
\item{n.var}{number of variables selected to use in each node split} 
\item{type}{classification}
\item{confusion}{confusion matrix} 
\item{call}{the original call to \code{forest}}
}
\description{
Run a Projection Pursuit forest
}
\examples{
tr.index <- train_fn(iris[, 5], 2/3)
train <- iris[sort(tr.index$id), 5:1 ]
test <- iris[-tr.index$id, 5:1 ]
ppfr.iris <- PPforest(train = train, testap = TRUE, test = test, m = 50, size.p = .9,
PPmethod = 'LDA', strata = TRUE)
}

