% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/PPplot.R
\name{PPplot}
\alias{PPplot}
\title{MDS}
\usage{
PPplot(ppfo, train, k)
}
\arguments{
  \item{ppfo}{a PPforest object} \item{train}{is a data
  frame with the training data with class in the first
  column} \item{k}{number of dimensions of the MDS layout}
}
\value{
MDS plot
}
\description{
MDS
}
\examples{
ppfr.iris <- PPforest( data = iris, size.tr=2/3, testap = TRUE,
 m = 500, size.p = .9, PPmethod = 'LDA', strata = TRUE, std=TRUE)
PPplot(ppfr.iris, train, k = 2)
PPplot(ppfr.iris, train, k = 3)
}

