% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/PPplot.R
\name{PPplot}
\alias{PPplot}
\title{Proximity matrix visualization}
\usage{
PPplot(ppfo, k)
}
\arguments{
\item{ppfo}{a PPforest object}

\item{k}{number of dimensions of the MDS layout}
}
\value{
MDS plot
}
\description{
MDS plot using proximity matrix inforation.
}
\examples{
ppfr.iris <- PPforest(data = iris[,5:1], size.tr = 2/3, testap = TRUE, m = 500, size.p = .9,
 PPmethod = 'LDA', strata = TRUE)
 PPplot(ppfr.iris, k = 2)
 PPplot(ppfr.iris, k = 3)
}

