% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/PPplot.R
\name{PPplot}
\alias{PPplot}
\title{MDS}
\usage{
PPplot(ppfo, train, k)
}
\arguments{
  \item{ppfo}{a PPforest object} \item{train}{is a data
  frame with the training data with class in the first
  column} \item{k}{number of dimensions of the MDS layout}
}
\value{
proximity matrix plot and MDS plot
}
\description{
MDS
}
\examples{
tr.index <- train_fn(iris[, 5], 2/3)
te.index <- as.vector(1:length(iris[, 5]))[!(1:length(iris[, 5]) \%in\% (sort(tr.index$id)))]
train <- iris[sort(tr.index$id), 5:1 ]
test <- iris[-tr.index$id, 5:1 ]
ppfr.iris <- PPforest( train = train, testap = TRUE, test = test, m = 500, size.p = .9, PPmethod = 'LDA', strata = TRUE)
PPplot(ppfr.iris, train, k = 2)
PPplot(ppfr.iris, train, k = 3)
}

