% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pproxy_plot.R
\name{pproxy_plot}
\alias{pproxy_plot}
\title{Proximity matrix visualization}
\usage{
pproxy_plot(ppf, type = "heat", k, inter = TRUE)
}
\arguments{
\item{ppf}{a PPforest object}

\item{type}{is MDS or heat}

\item{k}{number of dimensions of the MDS layout}

\item{inter}{condition if it is TRUE interactive plot}
}
\value{
proximity matrix plot, if type is MDS then a MDS plot using proximity matrix information is shown and if type is heat a heat map of the proximity matrix is shown
}
\description{
Proximity matrix visualization
}
\examples{
#leukemia data set with 2/3 observations used as training
pprf.leukemia <- PPforest(data = leukemia, class = 'Type',
size.tr = 2/3, m = 100, size.p = .5, PPmethod = 'PDA', strata = TRUE)
pproxy_plot(pprf.leukemia, type= "heat")
}

