% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/proximity_plot.R
\name{pproxy_plot}
\alias{pproxy_plot}
\title{Proximity matrix heat map plot}
\usage{
pproxy_plot(ppfo)
}
\arguments{
\item{ppfo}{a PPforest object}
}
\value{
proximity matrix plot
}
\description{
Proximity matrix heat map plot
}
\examples{
ppfr.iris <- PPforest(data = iris[,5:1], size.tr=2/3, m = 500, size.p = .9,
PPmethod = 'LDA', strata = TRUE)
pproxy_plot(ppfr.iris)
}

