% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pproxy_plot.R
\name{pproxy_plot}
\alias{pproxy_plot}
\title{Proximity matrix visualization}
\usage{
pproxy_plot(ppfo, type = "heat", k)
}
\arguments{
\item{ppfo}{a PPforest object}

\item{type}{is MDS or heat}

\item{k}{number of dimensions of the MDS layout}
}
\value{
proximity matrix plot, if type is MDS then a MDS plot using proximity matrix information is shown and if type is heat a heat map of the proximity matrix is shown
}
\description{
Proximity matrix visualization
}
\examples{
ppfr.iris <- PPforest(data = iris[,5:1], size.tr=2/3, m = 500, size.p = .9,
PPmethod = 'LDA', strata = TRUE)
pproxy_plot(ppfr.iris, type="heat")
pproxy_plot(ppfr.iris, type="MDS", k=2)
}

