% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pproxy_plot.R
\name{pproxy_plot}
\alias{pproxy_plot}
\title{Proximity matrix visualization}
\usage{
pproxy_plot(ppf, type = "heat", k)
}
\arguments{
\item{ppf}{a PPforest object}

\item{type}{is MDS or heat}

\item{k}{number of dimensions of the MDS layout}
}
\value{
proximity matrix plot, if type is MDS then a MDS plot using proximity matrix information is shown and if type is heat a heat map of the proximity matrix is shown
}
\description{
Proximity matrix visualization
}
\examples{
#leukemia data set with all the observations used as training
pprf.leukemia <- PPforest(data = leukemia, class = 'Type',
size.tr = 1, m = 70, size.p = .4, PPmethod = 'PDA', strata = TRUE)
pproxy_plot(pprf.leukemia, type= "MDS", k = 3)
}

