% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/PPtree_split.R
\name{PPtree_split}
\alias{PPtree_split}
\title{Projection pursuit classification tree with random variable selection in each split}
\usage{
PPtree_split(fr, data, PPmethod="LDA", weight=TRUE,
size.p=0.9, r=1, lambda=0.1, energy=0, maxiter=50000, ...)
}
\arguments{
  \item{fr}{is an object class formula, with the form
  response~terms} \item{data}{numeric data matrix without
  class information} \item{PPmethod}{index to use for
  projection pursuit: "LDA", "PDA", "Lr", "GINI", and
  "ENTROPY"} \item{weight}{flag in LDA, PDA and Lr index}
  \item{size.p}{proportion of variables used in each split}
  \item{r}{power in Lr index} \item{lambda}{tuning parameter
  in PDA index} \item{energy}{optimization parameter for
  projection pursuit} \item{maxiter}{number of maximum
  iteration} \item{...}{arguments to be passed to methods}
}
\value{
Tree.Struct Tree structure of projection pursuit
classification tree projbest.node 1-dim optimal projections
of each split node splitCutoff.node cutoff values of each
split node origclass original class origdata original data
}
\description{
Find tree structure using various projection pursuit indices
of classification in each split.
}
\examples{
data(iris)
Tree.result <- PPtree_split(as.formula('Species~.'), data=iris[,5:1], size.p=0.9)
Tree.result
}
\references{
Lee, YD, Cook, D., Park JW, and Lee, EK (2013) PPtree:
Projection pursuit classification tree, Electronic Journal
of Statistics, 7:1369-1386.
}
\keyword{tree}

