% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PPtree_split.R
\name{PPtree_split}
\alias{PPtree_split}
\title{Finding PP tree structure using various indices}
\usage{
PPtree_split(PPmethod,  i.class, i.data,size.p, weight = TRUE, ...)
}
\arguments{
  \item{PPmethod}{method for projection pursuit, LDA, PDA,
  Lp, Gini, Enp}

  \item{i.class}{class information}

  \item{i.data}{A training data without class information}

  \item{size.p}{proportion of sample variable in each
  split}

  \item{weight}{weight flag using in LDA index}
}
\value{
Tree.Struct Tree structure of PPtree result

Alpha.Keep 1D projections of each split

C.Keep spliting rules for each split
}
\description{
Find tree structure using various projection pursuit
indices of classification in each split.
}
\examples{
training<-train_fn(iris[,5],.9)
data1<-iris[,5:1]
Tree.result <- PPtree_split("LDA", data1[training,1], data1[training,2:5],size.p=0.9)
}

