% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/predict.PPforest.R
\name{predict.PPforest}
\alias{predict.PPforest}
\title{Obtain predicted class for new data using PPforest}
\usage{
\method{predict}{PPforest}(object, ...)
}
\arguments{
\item{object}{an object of class PPforest}

\item{...}{arguments to be passed to methods}
}
\value{
predicted values form PPforest
}
\description{
Vector with predicted values from a PPforest.
}
\examples{
pprf.iris <- PPforest(data = iris[,5:1], size.tr = .9, m = 500, size.p = .9,
PPmethod = 'PDA', strata = TRUE)
pr <- predict( object = pprf.iris, test.data = pprf.iris$test)
pr
}

