% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/predict.PPforest.R
\name{predict.PPforest}
\alias{predict.PPforest}
\title{Obtain predicted class for new data using PPforest}
\usage{
\method{predict}{PPforest}(object, test.data, Rule, true.class = NULL, ...)
}
\arguments{
\item{object}{an object of class PPforest}

\item{test.data}{the test dataset}

\item{Rule}{split rule \itemize{ \item{1: mean of two group means} \item{ 2: weighted mean of two group means - weight with group size} \item{3: weighted mean of two group means - weight with group sd} \item{ 4: weighted mean of two group means - weight with group se} \item{5: mean of two group medians} \item{ 6: weighted mean of two group medians - weight with group size} \item{ 7: weighted mean of two group median - weight with group IQR} \item{8: weighted mean of two group median - weight with group IQR and size} }}

\item{true.class}{true class of test dataset if available}

\item{...}{arguments to be passed to methods}
}
\value{
predicted values form PPforest
}
\description{
Vector with predicted values from a PPforest.
}
\examples{
pprf.leukemia <- PPforest(y = leukemia[, 1], x = leukemia[, -1],
 size.tr = 2/3, m = 500, size.p = .5, PPmethod = 'PDA', strata = TRUE)
pr.leukemia <- predict( object = pprf.leukemia, test.data = pprf.leukemia$test)
pr.leukemia
}

