% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/train_fn.R
\name{train_fn}
\alias{train_fn}
\title{Obtain stratified sample of the data}
\usage{
train_fn(y, size.p = 0.9)
}
\arguments{
\item{y}{input character vector with class of the data}

\item{size.p}{proportion of sample in each class group}
}
\value{
numeric vector  with the sample indexes for training data
}
\description{
Index id for training set, sample in each class with the same proportion.
}
\examples{
training.id <- train_fn(y = leukemia[,1], size.p = 0.9)
training.id
}

