% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/train_fn.R
\name{train_fn}
\alias{train_fn}
\title{Index id for training set, sample in each class with the same proportion.}
\usage{
train_fn(class, size.p)
}
\arguments{
\item{class}{input character vector with class of the data}

\item{size.p}{proportion of sample in each class group}
}
\value{
numeric vector giving number with the sample indexes for training
}
\description{
Index id for training set, sample in each class with the same proportion.
}
\examples{
training<-train_fn(class=iris[,5],size.p=0.9)
training
}

