% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tree_ppred.R
\name{tree_ppred}
\alias{tree_ppred}
\title{Obtain predicted class for new data using PPforest}
\usage{
tree_ppred(newdata, output.tree, ...)
}
\arguments{
\item{newdata}{are the complete data without the class variable we want to predict}

\item{output.tree}{trees classifiers from trees_pp function or PPforest object}

\item{...}{arguments to be passed to methods}
}
\value{
predicted values form PPforest
}
\description{
Vector with predicted values from a PPforest.
}
\examples{
data(iris)
training <- train_fn(class = iris[, 5], size.p = 2/3)
iris.b <- ppf_bootstrap( iris[training$id, 5:1], 500)
output <- trees_pp(iris.b, size.p = 0.9, PPmethod ='LDA')
pr <- tree_ppred( iris[-training$id, 1:4] , output)
pr
}

