% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_ppred2.R
\name{tree_ppred2}
\alias{tree_ppred2}
\title{Obtain predicted class for new data using PPforest}
\usage{
tree_ppred2(xnew, output.tree, parallel = FALSE, cores = 2)
}
\arguments{
\item{xnew}{data frame with explicative variables used to get new predicted values.}

\item{output.tree}{trees classifiers from trees_pp function or PPforest object}

\item{parallel}{if TRUE, apply function in parallel}

\item{cores}{The number of cores to use for parallel execution. By default is 2 cores.}
}
\value{
predicted values from PPforest
}
\description{
Vector with predicted values from a PPforest.
}
\examples{
leukemia.trees <- baggtree(data = leukemia, class = "Type", 
m =  70, PPmethod = 'PDA', lambda = .1, size.p = 0.4 ) 
pr <- tree_ppred2( xnew = leukemia[, -1] , leukemia.trees)
}

