% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/trees_pp.R
\name{trees_pp}
\alias{trees_pp}
\title{Projection persuit trees for bootstrap samples.}
\usage{
trees_pp(data.b, size.p = 0.9, PPmethod = "LDA", lambda = 0.1, ...)
}
\arguments{
  \item{data.b}{are the bootstrap samples. Class factor
  in the first column} \item{size.p}{proportion of random
  sample variables in each split.} \item{PPmethod}{to run
  the PPtree_plit function, options LDA or PDA, by default
  it is LDA.} \item{lambda}{a parameter for PDA index}
  \item{...}{arguments to be passed to methods}
}
\value{
data frame with tree_pp output for all the bootstraps
samples.
}
\description{
Projection persuit trees for bootstrap samples.
}
\examples{
training <- train_fn(iris[,5], .9)
data1 <- iris[training$id, 5:1]
iris.b <- bootstrap(data1, 500)
output1 <- trees_pp(iris.b, size.p=.9,  PPmethod = "LDA")
output2 <- trees_pp(iris.b, size.p=.9, PPmethod ='PDA', lambda=.1)
}

