% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/var_select.R
\name{var_select}
\alias{var_select}
\title{Index id for training set, sample in each class with the same proportion.}
\usage{
var_select(data, size.p)
}
\arguments{
  \item{data}{is a data frame without the class variable.}

  \item{size.p}{proportion of sample in each class group}
}
\value{
a vector giving the names of the selected variables
}
\description{
Index id for training set, sample in each class with the
same proportion.
}
\examples{
variables<-var_select(data=iris[,-1],size.p=0.9)
variables
}

