% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_select.R
\name{var_select}
\alias{var_select}
\title{Random selection of variables}
\usage{
var_select(data, size.p)
}
\arguments{
\item{data}{Data frame with the complete data set.}

\item{size.p}{Sample size proportion of variables randomly sampled as candidates at each split.}
}
\value{
a vector giving the id of the randomly sampled variables.
}
\description{
Index id for variables set, sample variables without replacement.
}
\examples{
variables <- var_select(data = leukemia,  size.p = 0.9)
variables
}

