% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/var_select.R
\name{var_select}
\alias{var_select}
\title{Random selection of variables}
\usage{
var_select(x, size.p)
}
\arguments{
\item{x}{is a data frame with explanatory variables.}

\item{size.p}{proportion of variables used in each split}
}
\value{
a vector giving the names of the selected variables
}
\description{
Index id for variables set, sample variables without replacement.
}
\examples{
variables <- var_select(x = leukemia[, -1], size.p = 0.9)
variables
}

