% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/var_select.R
\name{var_select}
\alias{var_select}
\title{Index id for variables set, sample variables without replacement with a sample proportion.}
\usage{
var_select(data, size.p)
}
\arguments{
  \item{data}{is a data frame without the class variable.}
  \item{size.p}{proportion of sample in each class group}
}
\value{
a vector giving the names of the selected variables
}
\description{
Index id for variables set, sample variables without
replacement with a sample proportion.
}
\examples{
variables <- var_select(data=iris[,-5], size.p=0.9)
variables
}

