% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vote_viz1.R
\name{vote_viz1}
\alias{vote_viz1}
\title{Vote matrix visualization}
\usage{
vote_viz1(data, class, ppf)
}
\arguments{
\item{data}{Data frame with the complete data set.}

\item{class}{A character with the name of the class variable.}

\item{ppf}{is a PPforest object}
}
\value{
A side-by-side jittered dotplot with the probability for each class to be classify in the corresponding class column.
Color by true observed class for each observation.
}
\description{
Vote matrix visualization
}
\examples{
#leukemia data set with all the observations used as training
pprf.leukemia <- PPforest(data = leukemia, class = "Type",
 size.tr = 1, m = 70, size.p = .4, PPmethod = 'PDA', strata = TRUE)
vote_viz1(leukemia,"Type",pprf.leukemia)
}

